# specs/002-full-stack-web/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo App API
  description: REST API for the Evolution of Todo - Phase II
  version: 2.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8000
    description: Development server

security:
  - BearerAuth: []

tags:
  - Health
  - Tasks

paths:
  /api/health:
    get:
      summary: Health check
      description: Verify the API is running
      tags: [Health]
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /api/tasks:
    get:
      summary: List all tasks
      description: Get all tasks for the authenticated user with optional filtering and sorting
      tags: [Tasks]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
          description: Filter by completion status
        - name: sort
          in: query
          schema:
            type: string
            enum: [created, title]
            default: created
          description: Sort field
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - JWT token missing or invalid
        '500':
          description: Internal server error

    post:
      summary: Create a new task
      description: Create a new task for the authenticated user
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid request (validation error)
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /api/tasks/{task_id}:
    get:
      summary: Get task details
      description: Get a specific task by ID (user must own the task)
      tags: [Tasks]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - task belongs to different user
        '404':
          description: Task not found
        '500':
          description: Internal server error

    put:
      summary: Update a task
      description: Update title, description, or completion status
      tags: [Tasks]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
        '500':
          description: Internal server error

    delete:
      summary: Delete a task
      description: Permanently delete a task
      tags: [Tasks]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
        '500':
          description: Internal server error

  /api/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion
      description: Toggle the completed status of a task
      tags: [Tasks]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Task status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
        '500':
          description: Internal server error

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth

  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Buy groceries
        description:
          type: string
          maxLength: 1000
          example: Get milk, eggs, and bread

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        completed:
          type: boolean

    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          example: Buy groceries
        description:
          type: string
          nullable: true
          example: Get milk, eggs, and bread
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2025-02-06T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-02-06T12:00:00Z"

    Error:
      type: object
      properties:
        detail:
          type: string
          example: Validation error

    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string